Display product images on frequent product cards
Migrate the `product_usage` table to include an `image_url` column and
update the database version. Store product image URLs when incrementing
product usage.

The `FrequentProductCard` widget now displays product images from the
stored URLs. If an image is not available or fails to load, a default
icon is shown. The usage count badge has been removed from the card.

Signed-off-by: Ales Drobysh <drobysh.ales@gmail.com>